move handler into subfolder of blockchains
